chore: Remove settings local storage logic (#6504)
