+++ frontend/src/context/settings-context.tsx
@@ -1,16 +1,9 @@
 import React from "react";
-import {
-  LATEST_SETTINGS_VERSION,
-  PostSettings,
-  Settings,
-  settingsAreUpToDate,
-} from "#/services/settings";
 import { useSettings } from "#/hooks/query/use-settings";
 import { useSaveSettings } from "#/hooks/mutation/use-save-settings";
+import { PostSettings, Settings } from "#/types/settings";
 
 interface SettingsContextType {
-  isUpToDate: boolean;
-  setIsUpToDate: (value: boolean) => void;
   saveUserSettings: (newSettings: Partial<PostSettings>) => Promise<void>;
   settings: Settings | undefined;
 }
@@ -27,8 +20,6 @@ export function SettingsProvider({ children }: SettingsProviderProps) {
   const { data: userSettings } = useSettings();
   const { mutateAsync: saveSettings } = useSaveSettings();
 
-  const [isUpToDate, setIsUpToDate] = React.useState(settingsAreUpToDate());
-
   const saveUserSettings = async (newSettings: Partial<PostSettings>) => {
     const updatedSettings: Partial<PostSettings> = {
       ...userSettings,
@@ -39,27 +30,15 @@ export function SettingsProvider({ children }: SettingsProviderProps) {
       delete updatedSettings.LLM_API_KEY;
     }
 
-    await saveSettings(updatedSettings, {
-      onSuccess: () => {
-        if (!isUpToDate) {
-          localStorage.setItem(
-            "SETTINGS_VERSION",
-            LATEST_SETTINGS_VERSION.toString(),
-          );
-          setIsUpToDate(true);
-        }
-      },
-    });
+    await saveSettings(updatedSettings);
   };
 
   const value = React.useMemo(
     () => ({
-      isUpToDate,
-      setIsUpToDate,
       saveUserSettings,
       settings: userSettings,
     }),
-    [isUpToDate, setIsUpToDate, saveUserSettings, userSettings],
+    [saveUserSettings, userSettings],
   );
 
   return <SettingsContext value={value}>{children}</SettingsContext>;
