+++ frontend/src/utils/settings-utils.ts
@@ -1,4 +1,4 @@
-import { Settings } from "#/services/settings";
+import { Settings } from "#/types/settings";
 
 const extractBasicFormData = (formData: FormData) => {
   const provider = formData.get("llm-provider")?.toString();
@@ -44,7 +44,7 @@ const extractAdvancedFormData = (formData: FormData) => {
   };
 };
 
-const extractSettings = (formData: FormData): Partial<Settings> => {
+export const extractSettings = (formData: FormData): Partial<Settings> => {
   const { LLM_MODEL, LLM_API_KEY, AGENT, LANGUAGE } =
     extractBasicFormData(formData);
 
@@ -65,12 +65,3 @@ const extractSettings = (formData: FormData): Partial<Settings> => {
     SECURITY_ANALYZER,
   };
 };
-
-const saveSettingsView = (view: "basic" | "advanced") => {
-  localStorage.setItem(
-    "use-advanced-options",
-    view === "advanced" ? "true" : "false",
-  );
-};
-
-export { extractSettings, saveSettingsView };
