+++ frontend/tests/conversation-panel.test.ts
@@ -34,7 +34,6 @@ test.beforeEach(async ({ page }) => {
   await page.evaluate(() => {
     localStorage.setItem("FEATURE_MULTI_CONVERSATION_UI", "true");
     localStorage.setItem("analytics-consent", "true");
-    localStorage.setItem("SETTINGS_VERSION", "5");
   });
 });
