+++ frontend/tests/settings.spec.ts
@@ -4,7 +4,6 @@ test.beforeEach(async ({ page }) => {
   await page.goto("/");
   await page.evaluate(() => {
     localStorage.setItem("analytics-consent", "true");
-    localStorage.setItem("SETTINGS_VERSION", "4");
   });
 });
